+++ source/frames/SDynamics.py
@@ -1,9 +1,17 @@
+import tkinter as tk
+import webbrowser
+
+import csv
+import tkcalendar
+import datetime
+from data import ScrumblesData
+from data import ScrumblesObjects
+
 from frames.SGraphs import *
 from frames.SLists import *
 
 from styling import styling as style
 from tkinter import ttk
-
 
 class SCalendar(tk.Frame):
     def __init__(self, controller):
