+++ source/testing/ScrumblesObjectsTest_userClass.py
@@ -1,5 +1,4 @@
-from data import ScrumblesData
-from frames import ScrumblesObjects
+from data import ScrumblesData, ScrumblesObjects
 
 
 def printUserDetails(user):
