+++ source/testing/testSprintDelete.py
@@ -1,5 +1,4 @@
-from data import DataBlock
-from frames import ScrumblesObjects
+from data import DataBlock, ScrumblesObjects
 from datetime import datetime
 import time
 import logging
