+++ source/views/mainView.py
@@ -1,6 +1,6 @@
 from tkinter import messagebox
-from frames import SPopMenu, ScrumblesObjects, Dialogs, ScrumblesFrames
-import listboxEventHandler
+from frames import SPopMenu, Dialogs, ScrumblesFrames, listboxEventHandler
+from data import ScrumblesObjects
 import logging
 import tkinter as tk
 
@@ -129,7 +129,7 @@ class mainView(tk.Frame):
         self.sprintDescriptionManager = ScrumblesFrames.SCardDescription(self, controller, sprintDynamicSources, sprintQueryType)
 
         itemDynamicSources = [self.fullBacklog.listbox, self.itemList.listbox, self.subItemList.listbox]
-        itemQueryType = ['Item', 'Item']
+        itemQueryType = ['Item', 'Item', 'Item']
         self.itemDescriptionManager = ScrumblesFrames.SCardDescription(self, controller, itemDynamicSources, itemQueryType)
 
         # To Prevent Duplicate Tkinter Events
